move util and filewatcher into separate file
