+++ lib/gpio.js
@@ -1,19 +1,17 @@
 var fs = require('fs');
-var util = require('util');
+var exists = fs.existsSync || path.exists;
 var path = require('path');
 var EventEmitter = require('events').EventEmitter;
-var exists = fs.existsSync || path.exists;
 
-var GPIO = {}, UTIL = {};
+var util = require('./util');
+var FileWatcher = require('./filewatcher');
 
-// default check interval when watching for changes
-var FILEWATCHER_INTERVAL = 100;
 
 /******************************************************************************
  * GPIO Exports
  *****************************************************************************/
 
-module.exports = GPIO;
+var GPIO = module.exports = {};
 
 GPIO.logging = false;
 
@@ -65,7 +63,7 @@ GPIO.export = function(pin, opts) {
  * @param {function(Error):void} callback
  */
 GPIO.close = function(pin, callback) {
-  UTIL.writeValue(pin, GPIO.PATH + 'unexport', function(err) {
+  util.writeValue(pin, GPIO.PATH + 'unexport', function(err) {
     if(typeof callback === 'function') {
       callback(err);
     }
@@ -85,171 +83,6 @@ GPIO.unexport = function(pin, callback) {
 
 
 /******************************************************************************
- * Utils
- *****************************************************************************/
-
-/**
- * log if error occurred
- * @param {object} err
- */
-UTIL.logError = function(err) {
-  if(GPIO.logging && err) {
-    console.error(err);
-  }
-};
-
-/**
- * log output of a call (if logging enabled)
- * @param {Object[]} arguments
- */
-UTIL.logMessage = function() {
-  if (exports.logging) {
-    console.log.apply(console, arguments);
-  }
-};
-
-/**
- * write value to file
- * @param {string} value
- * @param {string} filePath
- * @param {function} callback
- * @param {override} boolean
- */
-UTIL.writeValue = function(value, filePath, callback, override) {
-  if(typeof callback !== 'function') {
-    callback = UTIL.logError;
-  }
-
-  fs.writeFile(filePath, value, function(err) {
-    if(err && !override) {
-      err.path = filePath;
-      err.action = 'write';
-      UTIL.logError(err);
-    } else {
-      callback(err);
-    }
-  });
-};
-
-/**
- * read value from file
- * @param {string} filePath
- * @param {function} callback
- */
-UTIL.readValue = function(filePath, callback) {
-  fs.readFile(filePath, 'utf-8', function(err, data) {
-    if(err) {
-      err.path = filePath;
-      err.action = 'read';
-      UTIL.logError(err);
-    } else {
-      if(typeof callback === 'function') {
-        callback(data);
-      } else {
-        UTIL.logMessage('value: ', data);
-      }
-    }
-  });
-};
-
-/**
- * asynchronous attempt to open a GPIO pin
- * @param {number} pin
- * @param {function} callback
- */
-UTIL.openPin = function(pin, callback) {
-  if(exists(GPIO.PATH + 'gpio'+pin)) {
-    // already opened, close and open again
-    UTIL.logMessage('Pin already exported');
-    GPIO.close(pin, function() {
-      UTIL.openPin(pin, callback);
-    });
-  } else {
-    UTIL.logMessage('Exporting gpio' + pin);
-    UTIL.writeValue(pin, GPIO.PATH + 'export', function(err) {
-      // if there's an error when opening, close and repeat
-      if(err) {
-        GPIO.close(pin, function() {
-          UTIL.openPin(pin, callback);
-        });
-      } else if(typeof callback === 'function') {
-        callback();
-      }
-    }, true);
-  }
-};
-
-/**
- * check if file is writable
- * @param {string} filePath
- * @param {function} callback
- */
-UTIL.checkFileWritable = function(filePath, callback) {
-  fs.open(filePath, 'w', function(err, fd) {
-    if(err) {
-      callback(err);
-    } else {
-      fs.close(fd, function(err) {
-        callback(null);
-      });
-    }
-  });
-};
-
-
-/******************************************************************************
- * FileWatcher - watches for changes to a file
- * This custom watcher exists because fs.watch doesn't get fired due to file
- * never getting "accessed" if pin is set via hardware
- * @param {string} path
- * @param {number} interval
- * @param {function} callback
- *****************************************************************************/
-function FileWatcher(path, interval, callback) {
-  if(typeof callback === 'undefined') {
-    callback = interval;
-    interval = FILEWATCHER_INTERVAL;
-  }
-
-  if(typeof interval !== 'number' || typeof callback !== 'function') {
-    throw Error('Error initializing FileWatcher, check init arguments');
-  }
-
-  this.path       = path;
-  this.interval   = interval;
-  this.callback   = callback;
-  this.value      = null;
-  this.readTimer_ = null;
-
-  this.start();
-}
-
-/**
- * start watching file changes
- */
-FileWatcher.prototype.start = function() {
-  var self = this;
-  self.readTimer_ = setInterval(function() {
-    UTIL.readValue(self.path, function(val) {
-      if(self.value !== val) {
-        if(typeof self.value !== 'undefined') {
-          self.callback(val);
-        }
-        self.value = val;
-      }
-    });
-  }, self.interval);
-};
-
-/**
- * stop watching file changes
- */
-FileWatcher.prototype.stop = function() {
-  clearInterval(this.readTimer_);
-};
-
-
-/******************************************************************************
  * GPIO Pin
  * @param {object}   opts
  * @param {number}   opts.pin
@@ -263,10 +96,6 @@ function GPIOPin(opts, callback) {
   var pin = opts.pin;
   var direction = opts.direction;
   var interval = opts.interval;
-
-  if(typeof interval !== 'number') {
-    interval = FILEWATCHER_INTERVAL;
-  }
 
   if(typeof callback !== 'function') {
     callback = function() {};
@@ -287,15 +116,15 @@ function GPIOPin(opts, callback) {
     var makeAttempt = function() {
       attempts += 1;
 
-      UTIL.checkFileWritable(self.PATH.DIRECTION, function(err) {
+      util.checkFileWritable(self.PATH.DIRECTION, function(err) {
         if(err) {
           // Unable to gain write access
-          UTIL.logMessage('Could not write to pin: ' + err.code);
+          util.logMessage('Could not write to pin: ' + err.code);
           if(attempts <= 5) {
-            UTIL.logMessage('Trying again in 100ms');
+            util.logMessage('Trying again in 100ms');
             setTimeout(makeAttempt, 100);
           } else {
-            UTIL.logMessage('Failed to access pin after 5 attempts. Giving up.');
+            util.logMessage('Failed to access pin after 5 attempts. Giving up.');
             callback.call(self, 'Unable to open pin', self);
           }
 
@@ -324,14 +153,37 @@ function GPIOPin(opts, callback) {
 /**
  * Enable observer with EventEmitter class
  */
-util.inherits(GPIOPin, EventEmitter);
+require('util').inherits(GPIOPin, EventEmitter);
 
 /**
- * asyncronously open GPIO pin
+ * open GPIO pin
  * @param {function} callback
  */
 GPIOPin.prototype.open = function(callback) {
-  UTIL.openPin(this.pin, callback);
+  var self = this;
+
+  // already opened, close and open again
+  if(exists(GPIO.PATH + 'gpio'+self.pin)) {
+    util.logMessage('Pin already exported');
+    GPIO.close(self.pin, function() {
+      self.open(callback);
+    });
+    return;
+  }
+
+  // export pin
+  util.logMessage('Exporting gpio' + self.pin);
+  util.writeValue(self.pin, GPIO.PATH + 'export', function(err) {
+    // if there's an error when opening, close and repeat
+    if(err) {
+      GPIO.close(self.pin, function() {
+        self.open(callback);
+      });
+    } else if(typeof callback === 'function') {
+      callback();
+    }
+  }, true);
+
 };
 
 /**
@@ -380,7 +232,7 @@ GPIOPin.prototype.setDirection = function(direction, callback) {
   }
   self.direction = direction;
 
-  UTIL.logMessage('Setting direction "' + direction + '" on gpio' + self.pin);
+  util.logMessage('Setting direction "' + direction + '" on gpio' + self.pin);
 
   function watch () {
     if(direction === GPIO.DIRECTION.IN) {
@@ -393,6 +245,7 @@ GPIOPin.prototype.setDirection = function(direction, callback) {
           self.emit('valueChange', val);
           self.emit('change', val);
         });
+        self.interval = self.valueWatcher.interval;
       }
     } else {
       // if direction is OUT, try to clear the valueWatcher
@@ -403,16 +256,16 @@ GPIOPin.prototype.setDirection = function(direction, callback) {
     }
   }
 
-  UTIL.readValue(self.PATH.DIRECTION, function(currDir) {
+  util.readValue(self.PATH.DIRECTION, function(currDir) {
     var changedDirection = false;
     if(currDir.indexOf(direction) !== -1) {
-      UTIL.logMessage('Current direction is already ' + direction);
-      UTIL.logMessage('Attempting to set direction anyway.');
+      util.logMessage('Current direction is already ' + direction);
+      util.logMessage('Attempting to set direction anyway.');
     } else {
       changedDirection = true;
     }
 
-    UTIL.writeValue(direction, self.PATH.DIRECTION, function() {
+    util.writeValue(direction, self.PATH.DIRECTION, function() {
       watch();
 
       if(typeof callback === 'function') {
@@ -437,7 +290,7 @@ GPIOPin.prototype.get_ = function(callback) {
     return currentValue;
   }
 
-  UTIL.readValue(self.PATH.VALUE, function(value) {
+  util.readValue(self.PATH.VALUE, function(value) {
     value = parseInt(value, 10);
     if(value !== currentValue) {
       self.value = value;
@@ -475,7 +328,7 @@ GPIOPin.prototype.set = function(value, callback) {
   // if the value has changed; we don't have to rely on watching a file
   if(self.direction === GPIO.DIRECTION.OUT) {
     if(self.value !== value) {
-      UTIL.writeValue(value, self.PATH.VALUE, function() {
+      util.writeValue(value, self.PATH.VALUE, function() {
         self.value = value;
         self.emit('valueChange', value);
         self.emit('change', value);
